Grosse mise en forme des modules Chaque module est dans un fichier (.c et .h) séparé et utils ne comprend que des fonctions communes à tous les modules
