+++ main.c
@@ -1,4 +1,8 @@
-#include "fonction.h"
+#include "utils.h"
+#include "module_identite.h"
+#include "module_transcription.h"
+#include "module_traduction.h"
+#include "module_simi_pola.h"
 
 int main(){
 
@@ -21,49 +25,18 @@ int main(){
     printf("Malheureusement, ce service est encore indisponible \n");
   }
   else if (reponse == 2){
-    printf("Vous avez sélectionné : Transcription \n");
-    char path_input[100];
 
-    printf("quel est le nom du fichier que vous voulez lire \n");
-    scanf("%s", path_input); //nom du fichier test = test_fasta (fichier cours sans la première ligne)
-    //on ne peut mettre que le nom du fichier et non tout le chemin
-    printf("\n");
-
-    transcription(path_input);
+    transcription();
   }
   else if(reponse == 3){
-    printf("Vous avez sélectionné : Traduction \n");
-    char path_input[100];
 
-    printf("quel est le nom du fichier que vous voulez lire \n");
-    scanf("%s", path_input);
-    printf("\n");
-
-    traduction(path_input);
+    traduction();
   }
   else if(reponse == 4){
-    printf("Vous avez séletionné : Cacule du score d'identité entre deux séquence \n");
-    char path_input_1[100];
-    char path_input_2[100];
-    printf("vous avez besoin de deux fichiers \n");
-    printf("Quel est le nom du premier fichier \n");
-    scanf("%s", path_input_1);
-    printf("Quel est le nom du deuxième fichier \n");
-    scanf("%s", path_input_2);
-
-    identite(path_input_1,path_input_2);
+    identite();
   }
   else if(reponse == 5){
-    printf("Vous avez séletionné : Calcul du score de similarité de polarité entre deux séquences protéiques \n");
-    char path_input_1[100];
-    char path_input_2[100];
-    printf("Vous avez besoin de deux fichiers \n");
-    printf("Quel est le nom du premier fichier \n");
-    scanf("%s", path_input_1);
-    printf("Quel est le nom du deuxième fichier \n");
-    scanf("%s", path_input_2);
-
-    simi_pola(path_input_1,path_input_2);
+    simi_pola();
   }
   else if(reponse == 6){
     printf("Vous avez sélectionné : Recherche séquence consensus issue d’un résultat d’alignement multiple \n");
