+++ module_traduction.h
@@ -1,0 +1,3 @@
+#include "utils.h"
+
+void traduction();
